test-requirements.txt
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
bcrypt==3.1.7
beautifulsoup4==4.7.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
decorator==4.4.0
Flask==1.1.1
Flask-MySQLdb==0.2.0
http-ece==1.1.0
idna==2.8
importlib-metadata==0.22
itsdangerous==1.1.0
Jinja2==2.10.1
markovify==0.7.1
MarkupSafe==1.1.1
Mastodon.py==1.4.6
more-itertools==7.2.0
mysqlclient==1.4.4
packaging==19.1
pluggy==0.13.0
py==1.8.0
pycparser==2.19
pyparsing==2.4.2
pytest==5.1.2
python-dateutil==2.8.0
python-magic==0.4.15
pytz==2019.2
requests==2.22.0
six==1.12.0
soupsieve==1.9.3
Unidecode==1.1.1
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.6
zipp==0.6.0
